+++ library/file/src/nativeMain/kotlin/io/matthewnelson/kmp/file/internal/NativeFs.kt
@@ -17,7 +17,9 @@
 
 package io.matthewnelson.kmp.file.internal
 
+import io.matthewnelson.kmp.file.File
 import io.matthewnelson.kmp.file.IOException
+import io.matthewnelson.kmp.file.OpenMode
 import kotlinx.cinterop.ByteVar
 import kotlinx.cinterop.CPointer
 import kotlinx.cinterop.ExperimentalForeignApi
@@ -47,6 +49,16 @@ internal expect fun fs_platform_mkdir(path: Path): Int
 internal expect inline fun MemScope.fs_platform_file_size(
     path: Path,
 ): Long
+
+@ExperimentalForeignApi
+@Throws(IllegalArgumentException::class, IOException::class)
+internal expect inline fun File.fs_platform_fopen(
+    flags: Int,
+    format: String,
+    b: Boolean,
+    e: Boolean,
+    mode: OpenMode,
+): CPointer<FILE>
 
 @ExperimentalForeignApi
 internal expect inline fun fs_platform_fread(
